io.c: Fix rebuffering issue on win32 (fixing #235) webserver.c: Fix write error log messages.  Possible fix for crash on mipsel (uninitialized heap issue).

+++ src/webserver.c
@@ -1203,7 +1203,7 @@ int ws_writefd(WS_CONNINFO *pwsc, char *fmt, ...) {
     }
 
     WS_EXIT();
-    return len;
+    return (int)len;
 }
 
 /**
